feat: move page-switcher style bb-344
