+++ apps/mobile/src/libs/components/components.tsx
@@ -16,6 +16,7 @@ export { LoaderWrapper } from "./loader-wrapper/loader-wrapper";
 export { Logo } from "./logo/logo";
 export { MultipleCheckboxInput } from "./multiple-checkbox-input/multiple-checkbox-input";
 export { MultipleSliderInput } from "./multiple-slider-input/multiple-slider-input";
+export { PageSwitcher } from "./page-switcher/page-switcher";
 export { Planet } from "./planet/planet";
 export { ProgressBar } from "./progress-bar/progress-bar";
 export { RadioGroup } from "./radio-group/radio-group";
